Stage 3 continuation, bugfixes, code cleanup, better exception handling

+++ src/main/java/org/training/dto/player/PlayerUpdateRequest.java
@@ -9,6 +9,6 @@ public class PlayerUpdateRequest {
 
     private String email;
 
-    private Integer eloRating;
+    private Double eloRating;
 
 }
